SA-1957: Renamed file
